@@ -1,90 +1,92 @@
 {
   "name": "ooni-explorer",
-  "version": "2.0.9",
+  "version": "2.4.4",
   "author": "Open Observatory of Network Interference (OONI) <contact@openobservatory.org>",
   "license": "BSD-3-Clause",
   "main": "index.js",
   "repository": "https://github.com/ooni/explorer",
   "dependencies": {
-    "@babel/cli": "7.4.4",
     "@babel/core": "7.4.4",
-    "@datapunt/matomo-tracker-js": "^0.0.15",
+    "@datapunt/matomo-tracker-react": "^0.3.1",
+    "@nivo/bar": "^0.62.0",
+    "@nivo/calendar": "^0.62.0",
+    "@nivo/heatmap": "^0.62.0",
+    "@rebass/forms": "^4.0.6",
     "@sentry/browser": "^5.0.8",
     "@sentry/integrations": "^5.0.8",
     "@sentry/node": "^5.0.8",
-    "@zeit/next-css": "1.0.1",
     "@zeit/next-source-maps": "0.0.3",
-    "axios": "^0.18.1",
+    "axios": "^0.21.1",
     "babel-plugin-inline-react-svg": "1.1.0",
     "babel-plugin-styled-components": "^1.10.7",
-    "babel-plugin-transform-define": "1.3.1",
     "buffer-from": "^1.1.1",
     "country-util": "0.1.0",
-    "dotenv": "^4.0.0",
+    "deepmerge": "^4.2.2",
     "express": "^4.14.0",
-    "file-loader": "3.0.1",
     "flag-icon-css": "^2.9.0",
+    "fontsource-fira-sans": "^3.0.5",
     "lodash.debounce": "^4.0.8",
-    "lodash.random": "^3.2.0",
-    "markdown-to-jsx": "^6.9.4",
+    "markdown-to-jsx": "^6.11.4",
     "moment": "^2.20.1",
-    "next": "9.3.2",
-    "node-html-to-image": "^2.1.1",
+    "next": "^10.0.1",
     "nprogress": "^0.2.0",
-    "ooni-components": "^0.3.0",
+    "ooni-components": "^0.4.1",
     "pretty-ms": "^3.2.0",
     "prop-types": "15.7.2",
-    "react": "^16.12.0",
+    "react": "16.14.0",
     "react-content-loader": "^4.2.2",
     "react-datetime": "2.16.3",
-    "react-dom": "^16.12.0",
+    "react-dom": "16.14.0",
+    "react-hook-form": "^6.12.2",
     "react-icons": "2.2.7",
     "react-intl": "3.3.2",
     "react-json-view": "^1.16.0",
-    "react-no-ssr": "^1.1.0",
+    "react-portal-tooltip": "^2.4.7",
     "react-sticky": "^6.0.3",
     "react-table": "^7.0.0-rc.15",
-    "rebass": "2.3.3",
+    "rebass": "4.0.7",
     "sentry-testkit": "^2.2.0",
     "serve-favicon": "^2.5.0",
     "styled-components": "^4.4.1",
-    "three": "^0.86.0",
-    "typeface-fira-sans": "^0.0.54",
+    "swr": "^0.2.3",
     "url": "^0.11.0",
-    "url-loader": "1.1.2",
     "use-clipboard-copy": "^0.1.2",
     "victory": "^32.0.2"
   },
   "resolutions": {
     "hoist-non-react-statics": "3.0.1",
-    "rebass": "2.3.3",
+    "rebass": "4.0.7",
     "styled-components": "^4.4.1",
-    "styled-system": "3.1.6",
+    "styled-system": "5.1.5",
     "react-icons": "2.2.7",
     "js-yaml": "^3.13.1",
-    "serialize-javascript": "^2.1.1",
+    "serialize-javascript": "^3.1.0",
     "https-proxy-agent": "^2.2.3",
-    "kind-of": "^6.0.3"
+    "kind-of": "^6.0.3",
+    "object-path": "^0.11.5",
+    "node-fetch": "^2.6.1",
+    "http-proxy": "^1.18.1"
   },
   "devDependencies": {
     "babel-plugin-react-intl": "^3.2.0",
-    "cypress": "^4.1.0",
-    "eslint": "^4.19.1",
-    "eslint-plugin-react": "^7.7.0",
+    "cypress": "^5.2.0",
+    "eslint": "^7.20.0",
+    "eslint-plugin-react": "^7.22.0",
+    "glob": "^7.1.6",
     "mustache": "^2.3.0",
     "puppeteer": "^5.0.0",
     "start-server-and-test": "^1.10.8"
   },
   "scripts": {
-    "dev": "NODE_ENV=development MEASUREMENTS_URL=http://api.ooni.io SENTRY_DSN=https://49af7fff247c445b9a7c98ee21ddfd2f@sentry.io/1427510 node index.js",
+    "dev": "NODE_ENV=development MEASUREMENTS_URL=https://api.ooni.io SENTRY_DSN=https://49af7fff247c445b9a7c98ee21ddfd2f@sentry.io/1427510 node index.js",
     "build-translations": "node ./scripts/build-translations.js",
     "extract-i18n-keys": "yarn build && node ./scripts/extract-react-intl-keys.js",
     "build": "next build",
     "export": "next export",
-    "lint": "eslint --ignore-path .gitignore --ignore-pattern components/vendor .",
-    "test:e2e": "start-server-and-test dev http://localhost:3100 cypress:run",
+    "lint": "eslint --ignore-path .gitignore .",
+    "test:e2e": "yarn build && start-server-and-test start http://localhost:3100 cypress:run",
     "cypress:run": "cypress run",
     "test": "echo nothing interesting",
-    "start": "node index.js"
+    "start": "MEASUREMENTS_URL=https://api.ooni.io node index.js"
   }
 }
