@@ -57,9 +57,7 @@ const NetworkRow = ({ asn, app }) => {
               </strong>
             </Box>
             <Box>
-              <Link
-                href={linkToMeasurements}
-              >
+              <Link href={linkToMeasurements} color='blue7'>
                 <FormattedMessage id='Country.Websites.URLCharts.ExploreMoreMeasurements' />
               </Link>
             </Box>
@@ -134,7 +132,7 @@ class AppsStatRow extends React.Component {
         </Flex>
         {(visibleNetworks < totalNetworks) &&
           <Flex justifyContent='center'>
-            <Link color='blue7' href='javascript:void(0)' onClick={() => this.showMore()}>
+            <Link color='blue7' href='#' onClick={(e) => {e.preventDefault(); this.showMore()}}>
               <FormattedMessage id='Country.Apps.Button.ShowMore' />
             </Link>
           </Flex>
@@ -167,7 +165,7 @@ class AppsStatRow extends React.Component {
             <CollapseTrigger
               size={36}
               bg={theme.colors.gray3}
-              isOpen={!minimized}
+              open={!minimized}
               onClick={this.toggleMinimize}
             />
           </Box>
