@@ -1,6 +1,6 @@
 import React from 'react'
 
-import SvgFlagZZ from '../../static/flags/1x1/zz.svg'
+import SvgFlagZZ from '../../public/static/flags/1x1/zz.svg'
 
 export const FlagZZ = ({size = 512}) => (
   <SvgFlagZZ viewBox='0 0 512 512' height={size} width={size}/>
