@@ -1,7 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Flex, Box } from 'ooni-components'
-import { Text } from 'rebass'
+import { Flex, Box, Text } from 'ooni-components'
 
 import HighlightBox from './HighlightBox'
 
@@ -37,12 +36,18 @@ const HighlightSection = ({
 }
 
 HighlightSection.propTypes = {
-  title: PropTypes.string.isRequired,
-  description: PropTypes.string,
+  title: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.element
+  ]),
+  description: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.element
+  ]),
   highlights: PropTypes.arrayOf(PropTypes.shape({
     countryCode: PropTypes.string.isRequired,
     countryName: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
+    title: PropTypes.string,
     text: PropTypes.string,
     report: PropTypes.string,
     explore: PropTypes.string
