@@ -1,4 +1,4 @@
-/* global require */
+/* global process */
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
@@ -10,34 +10,36 @@ import {
   Link,
   theme
 } from 'ooni-components'
-
-import NoSSR from 'react-no-ssr'
-import { useIntl } from 'react-intl'
+import dynamic from 'next/dynamic'
+import { FormattedMessage, useIntl } from 'react-intl'
+import { useRouter } from 'next/router'
 
 import { DetailsBoxTable, DetailsBox } from './DetailsBox'
 
-// We wrap the json viewer so that we can render it only in client side rendering
-class JsonViewer extends React.Component {
-  render() {
-    const ReactJson = require('react-json-view').default
-    const {
-      src
-    } = this.props
-    const StyledReactJsonContainer = styled.div`
-      .string-value {
-        text-overflow: ellipsis;
-        max-width: 800px;
-        overflow: hidden;
-        display: inline-block;
-      }
-    `
-    return (
-      <StyledReactJsonContainer>
-        <ReactJson collapsed={1} src={src} />
-      </StyledReactJsonContainer>
-    )
-  }
+const LoadingRawData = (props) => {
+  return (<Box fontSize={1}>Loading</Box>)
 }
+
+const ReactJson = dynamic(
+  () => import('react-json-view'),
+  { ssr: false, loading: LoadingRawData }
+
+)
+
+const StyledReactJsonContainer = styled.div`
+  .string-value {
+    text-overflow: ellipsis;
+    max-width: 800px;
+    overflow: hidden;
+    display: inline-block;
+  }
+`
+
+const JsonViewer = ({ src }) => (
+  <StyledReactJsonContainer>
+    <ReactJson collapsed={1} src={src} />
+  </StyledReactJsonContainer>
+)
 
 JsonViewer.propTypes = {
   src: PropTypes.object.isRequired
@@ -45,42 +47,56 @@ JsonViewer.propTypes = {
 
 const CommonDetails = ({
   measurement,
-  measurementURL
+  reportId
 }) => {
   const {
-    report_id,
+    software_name,
     software_version,
     annotations,
-  } = measurement
+  } = measurement ?? {}
+
+  const { query } = useRouter()
+  const queryString = new URLSearchParams(query)
+  const rawMsmtDownloadURL = `${process.env.MEASUREMENTS_URL}/api/v1/raw_measurement?${queryString}`
+
   const intl = useIntl()
+  const unavailable = intl.formatMessage({ id: 'Measurement.CommonDetails.Value.Unavailable' })
 
-  let engine_version = 'none',
-    platform = 'unknown'
+  let engine = unavailable
+  let platform = unavailable
 
-  if (annotations && annotations.engine_version) {
-    engine_version = annotations.engine_version
+  if (annotations && annotations.engine_name) {
+    engine = annotations.engine_name
+
+    if (annotations.engine_version) {
+      engine = `${engine} (${annotations.engine_version})`
+    }
   }
+
   if (annotations && annotations.platform) {
     platform = annotations.platform
   }
 
-  const downloadFilename = `ooni-measurement-${report_id}.json`
+  let software = software_name ?? unavailable
+  software += software_version ? ` (${software_version})` : ''
+
+  const downloadFilename = `ooni-measurement-${reportId}.json`
   const items = [
     {
       label: intl.formatMessage({ id: 'Measurement.CommonDetails.Label.MsmtID' }),
-      value: report_id
+      value: reportId ?? unavailable
     },
     {
       label: intl.formatMessage({ id: 'Measurement.CommonDetails.Label.Platform' }),
-      value: platform ? platform : 'unknown'
+      value: platform
     },
     {
-      label: intl.formatMessage({ id: 'Measurement.CommonDetails.Label.SoftwareVersion' }),
-      value: software_version
+      label: intl.formatMessage({ id: 'Measurement.CommonDetails.Label.Software' }),
+      value: software
     },
     {
-      label: intl.formatMessage({ id: 'Measurement.CommonDetails.Label.MKVersion' }),
-      value: engine_version
+      label: intl.formatMessage({ id: 'Measurement.CommonDetails.Label.Engine' }),
+      value: engine ?? intl.formatMessage({ id: 'Measurement.CommonDetails.Value.Unavailable' })
     }
   ]
   return (
@@ -102,22 +118,27 @@ const CommonDetails = ({
                 <Heading h={4}>{intl.formatMessage({ id: 'Measurement.CommonDetails.RawMeasurement.Heading' })}</Heading>
               </Box>
               <Box >
-                <Link color='blue7' href={measurementURL} download={downloadFilename}>
+                <Link color='blue7' href={rawMsmtDownloadURL} download={downloadFilename}>
                   <Button
                     onClick={(e) => e.stopPropagation()}
                     fontSize={13}
                     mx={3}
-                    px={3}>{intl.formatMessage({ id: 'Measurement.CommonDetails.RawMeasurement.Download' })}</Button>
+                    px={3}
+                  >
+                    {intl.formatMessage({ id: 'Measurement.CommonDetails.RawMeasurement.Download' })}
+                  </Button>
                 </Link>
               </Box>
             </Flex>
           }
           content={
-            <Flex bg='WHITE' p={3}>
-              <NoSSR>
+            measurement && typeof measurement === 'object' ? (
+              <Flex bg='WHITE' p={3}>
                 <JsonViewer src={measurement} />
-              </NoSSR>
-            </Flex>
+              </Flex>
+            ) : (
+              <FormattedMessage id='Measurement.CommonDetails.RawMeasurement.Unavailable' />
+            )
           }
         />
       </Flex>
@@ -126,8 +147,8 @@ const CommonDetails = ({
 }
 
 CommonDetails.propTypes = {
-  measurement: PropTypes.object.isRequired,
-  measurementURL: PropTypes.string.isRequired
+  measurement: PropTypes.object,
+  reportId: PropTypes.string
 }
 
 export default CommonDetails
