@@ -13,40 +13,37 @@ const PerformanceDetails = ({
   timeouts
 }) => {
   const intl = useIntl()
-  let items = [
-    {
-      label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.AvgPing' }),
-      value: averagePing.toString() + ' ms'
-    },
-    {
-      label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.MaxPing' }),
-      value: `${isNdt7 ? '~' : ''}${maxPing.toString()} ms`
-    },
-    {
-      label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.MSS' }),
-      value: mss.toString()
-    },
-    {
-      label: isNdt7 ? intl.formatMessage({
-        id: 'Measurement.Details.Performance.Label.RetransmitRate',
-        defaultMessage: 'Retransmit Rate'
-      }): intl.formatMessage({ id: 'Measurement.Details.Performance.Label.PktLoss' }),
-      value: packetLoss.toString() + '%'
-    },
-  ]
+  let items = []
+  averagePing && items.push({
+    label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.AvgPing' }),
+    value: `${averagePing} ms`
+  })
+  maxPing && items.push({
+    label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.MaxPing' }),
+    value: `${isNdt7 ? '~' : ''}${maxPing} ms`
+  })
+  mss && items.push({
+    label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.MSS' }),
+    value: `${mss}`
+  })
+
+  packetLoss != undefined && items.push({
+    label: isNdt7 ? intl.formatMessage({
+      id: 'Measurement.Details.Performance.Label.RetransmitRate'
+    }): intl.formatMessage({ id: 'Measurement.Details.Performance.Label.PktLoss' }),
+    value: `${packetLoss}%`
+  })
 
   //Only add outOfOrder and timeouts if NDT4/5 measurement
-  if(!isNdt7){
-    items = items.concat([
-      {
-        label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.OutOfOrder' }),
-        value: outOfOrder.toString() + '%'
-      },
-      {
-        label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.Timeouts' }),
-        value: timeouts.toString()
-      }
-    ])
+  if(!isNdt7) {
+    outOfOrder != undefined && items.push({
+      label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.OutOfOrder' }),
+      value: outOfOrder.toString() + '%'
+    })
+    timeouts != undefined && items.push({
+      label: intl.formatMessage({ id: 'Measurement.Details.Performance.Label.Timeouts' }),
+      value: timeouts.toString()
+    })
   }
 
   return (
