@@ -1,15 +1,10 @@
-import React from 'react'
 import PropTypes from 'prop-types'
-import { FormattedMessage } from 'react-intl'
 
-const DefaultTestDetails = ({ measurement, render }) => (
-  render({
-    statusLabel: <FormattedMessage id='Measurement.Hero.Status.Default' />,
-  })
+const DefaultTestDetails = ({ render }) => (
+  render({})
 )
 
 DefaultTestDetails.propTypes = {
-  measurement: PropTypes.object,
   render: PropTypes.func
 }
 
