@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import {
   Text
 } from 'ooni-components'
-import { FormattedMessage } from 'react-intl'
+import { FormattedMessage, defineMessages } from 'react-intl'
 
 const HttpInvalidRequestLineDetails = ({ measurement, render }) => {
   const testKeys = measurement.test_keys
@@ -13,6 +13,18 @@ const HttpInvalidRequestLineDetails = ({ measurement, render }) => {
 
   const sentMessages = testKeys.sent
   const receivedMessages = testKeys.received
+
+  const messages = defineMessages({
+    middleboxes: {
+      id: 'Measurement.Metadata.HTTPInvalidReqLine.Middleboxes',
+      defaultMessage: 'Network traffic manipulation was detected in {country}'
+    },
+    noMiddleboxes: {
+      id: 'Measurement.Metadata.HTTPInvalidReqLine.NoMiddleboxes',
+      defaultMessage: 'Network traffic manipulation was not detected in {country}'
+    }
+  })
+
   return (
     render({
       status: isAnomaly ? 'anomaly' : 'reachable',
@@ -22,6 +34,10 @@ const HttpInvalidRequestLineDetails = ({ measurement, render }) => {
       summaryText: isAnomaly
         ? 'Measurement.HTTPInvalidReqLine.MiddleboxesDetected.SummaryText'
         : 'Measurement.HTTPInvalidReqLine.NoMiddleBoxes.SummaryText',
+      headMetadata: {
+        message: isAnomaly ? messages.middleboxes : messages.noMiddleboxes,
+        formatted: false
+      },
       details: (
         <div>
           {/*<Text>isAnomaly: {isAnomaly.toString()}</Text>
