@@ -11,6 +11,7 @@ import {
 import { FormattedMessage } from 'react-intl'
 import MdPhoneAndroid from 'react-icons/lib/md/phone-android'
 import MdWebAsset from 'react-icons/lib/md/web-asset'
+import { defineMessages } from 'react-intl'
 
 import AccessPointStatus from '../AccessPointStatus'
 import { DetailsBox } from '../DetailsBox'
@@ -24,26 +25,41 @@ const TelegramDetails = ({ measurement, render }) => {
     tcp_connect
   } = testKeys
 
+  const message = defineMessages({
+    reachable: {
+      id: 'Measurement.Metadata.Telegram.Reachable',
+      defaultMessage: 'Telegram was reachable in {country}'
+    },
+    unReachable: {
+      id: 'Measurement.Metadata.Telegram.UnReachable',
+      defaultMessage: 'Telegram was NOT reachable in {country}'
+    }
+  })
+
   let telegramWebOK = true
   let telegramDesktopOK = true
   let anomaly = false
   let hint = <FormattedMessage id='Measurement.Status.Hint.Telegram.Reachable' />
   let summaryText = 'Measurement.Details.SummaryText.Telegram.Reachable'
+  let headMetadata = message.reachable
 
   if (telegram_web_status === 'blocked') {
     telegramWebOK = false
     summaryText = 'Measurement.Details.SummaryText.Telegram.AppFailure'
+    headMetadata = message.unReachable
   }
 
   if (telegram_tcp_blocking === true || telegram_http_blocking === true) {
     telegramDesktopOK = false
     summaryText = 'Measurement.Details.SummaryText.Telegram.DesktopFailure'
+    headMetadata = message.unReachable
   }
 
   if (!telegramWebOK || !telegramDesktopOK) {
     anomaly = true
     hint = <FormattedMessage id='Measurement.Status.Hint.Telegram.Blocked' />
     summaryText = 'Measurement.Details.SummaryText.Telegram.DesktopAndAppFailure'
+    headMetadata = message.unReachable
   }
 
   return (
@@ -51,6 +67,10 @@ const TelegramDetails = ({ measurement, render }) => {
       status: anomaly ? 'anomaly': 'reachable',
       statusInfo: hint,
       summaryText: summaryText,
+      headMetadata: {
+        message: headMetadata,
+        formatted: false
+      },
       details: (
         <React.Fragment>
           <Container>
@@ -73,28 +93,28 @@ const TelegramDetails = ({ measurement, render }) => {
                       />
                     </Box>
                   </Flex>
-                  {tcp_connect && tcp_connect.length > 0 &&
-                    <div>
-                      <Heading h={4}> <FormattedMessage id='Measurement.Details.Telegram.Endpoint.Status.Heading' /> </Heading>
-                      {tcp_connect.map((connection, index) => (
-                        <Flex key={index}>
-                          <Box>
-                            <Text>
-                              {connection.status.failure &&
-                                <FormattedMessage id="Measurement.Details.Telegram.Endpoint.ConnectionTo.Failed"
-                                  values={{ destination: <strong> {connection.ip}:{connection.port} </strong> }}
-                                />
-                              }
-                              {connection.status.success &&
-                                <FormattedMessage id="Measurement.Details.Telegram.Endpoint.ConnectionTo.Successful"
-                                  values={{ destination: <strong> {connection.ip}:{connection.port} </strong> }}
-                                />
-                              }
-                            </Text>
-                          </Box>
-                        </Flex>
-                      ))}
-                    </div>
+                  {Array.isArray(tcp_connect) && tcp_connect.length > 0 &&
+                  <div>
+                    <Heading h={4}> <FormattedMessage id='Measurement.Details.Telegram.Endpoint.Status.Heading' /> </Heading>
+                    {tcp_connect.map((connection, index) => (
+                      <Flex key={index}>
+                        <Box>
+                          <Text>
+                            {connection.status.failure &&
+                            <FormattedMessage id="Measurement.Details.Telegram.Endpoint.ConnectionTo.Failed"
+                              values={{ destination: <strong> {connection.ip}:{connection.port} </strong> }}
+                            />
+                            }
+                            {connection.status.success &&
+                              <FormattedMessage id="Measurement.Details.Telegram.Endpoint.ConnectionTo.Successful"
+                                values={{ destination: <strong> {connection.ip}:{connection.port} </strong> }}
+                              />
+                            }
+                          </Text>
+                        </Box>
+                      </Flex>
+                    ))}
+                  </div>
                   }
                 </React.Fragment>
               } />
