@@ -6,7 +6,7 @@ import {
   Flex,
   Box
 } from 'ooni-components'
-import { FormattedMessage } from 'react-intl'
+import { FormattedMessage, defineMessages } from 'react-intl'
 import MdPhoneAndroid from 'react-icons/lib/md/phone-android'
 import MdWebAsset from 'react-icons/lib/md/web-asset'
 import MdPersonAdd from 'react-icons/lib/md/person-add'
@@ -35,6 +35,17 @@ const WhatsAppDetails = ({ isAnomaly, scores, measurement, render }) => {
     webAccessible = testKeys.whatsapp_web_status === 'ok'
   }
 
+  const messages = defineMessages({
+    reachable: {
+      id: 'Measurement.Metadata.Whatsapp.Reachable',
+      defaultMessage: 'WhatsApp was reachable in {country}'
+    },
+    unReachable: {
+      id: 'Measurement.Metadata.Whatsapp.UnReachable',
+      defaultMessage: 'WhatsApp was likely blocked in {country}'
+    }
+  })
+
   let status = 'reachable'
   let info = <FormattedMessage id='Measurement.Details.Hint.WhatsApp.Reachable' />
   let summaryText = 'Measurement.Details.SummaryText.WhatsApp.Reachable'
@@ -51,16 +62,16 @@ const WhatsAppDetails = ({ isAnomaly, scores, measurement, render }) => {
     }
   }
 
-
   return render({
     status: status,
     statusInfo: info,
     summaryText: summaryText,
+    headMetadata: {
+      message: isAnomaly ? messages.unReachable : messages.reachable,
+      formatted: false
+    },
     details: (
       <React.Fragment>
-        { !scores.analysis && <BugBox>
-        <Text>Attention! This test result may not be accurate as it may be affected by a <a href="https://github.com/ooni/explorer/issues/406">known data processing issue</a> that we are working on fixing. If in doubt on how to interpret this particular result, feel free to <a href="https://ooni.org/about">reach out to the OONI team</a>.</Text>
-        </BugBox>}
         <Box width={1/2}>
           <Flex>
             <Box width={1/3}>
@@ -86,7 +97,7 @@ const WhatsAppDetails = ({ isAnomaly, scores, measurement, render }) => {
             </Box>
           </Flex>
         </Box>
-        {tcp_connect && tcp_connect.length > 0 &&
+        {Array.isArray(tcp_connect) && tcp_connect.length > 0 &&
           <React.Fragment>
             <Heading h={4}> <FormattedMessage id='Measurement.Details.WhatsApp.Endpoint.Status.Heading' /> </Heading>
             {tcp_connect.map((connection, index) => (
