@@ -1,5 +1,5 @@
 import React from 'react'
-import Document, { Head, Main, NextScript } from 'next/document'
+import Document, { Html, Head, Main, NextScript } from 'next/document'
 import { ServerStyleSheet } from 'styled-components'
 
 export default class MyDocument extends Document {
@@ -12,9 +12,8 @@ export default class MyDocument extends Document {
 
   render () {
     return (
-      <html>
+      <Html lang='en'>
         <Head>
-          <link rel='stylesheet' href='/_next/static/style.css' />
           {this.props.styleTags}
         </Head>
         <body>
@@ -22,7 +21,7 @@ export default class MyDocument extends Document {
           <script src='/static/lang/translations.js' />
           <NextScript />
         </body>
-      </html>
+      </Html>
     )
   }
 }
