@@ -1,8 +1,9 @@
 /* global process */
 import React from 'react'
+import PropTypes from 'prop-types'
 import Head from 'next/head'
 import NLink from 'next/link'
-import Router from 'next/router';
+import Router from 'next/router'
 import FormattedMarkdown from '../components/FormattedMarkdown'
 import styled from 'styled-components'
 import axios from 'axios'
@@ -14,19 +15,17 @@ import {
   Container,
   Button,
   Heading,
+  Text
 } from 'ooni-components'
-import { Text } from 'rebass'
 
 import Layout from '../components/Layout'
 import NavBar from '../components/NavBar'
 import { toCompactNumberUnit } from '../utils'
 import HighlightSection from '../components/landing/HighlightsSection'
 import highlightContent from '../components/landing/highlights.json'
-import { Flag } from '../components/Flag'
 import CoverageChart from '../components/landing/Stats'
 
 const HeroUnit = styled.div`
-  background: url(/static/images/world-dots.svg);
   background: linear-gradient(
     319.33deg,
     ${props => props.theme.colors.blue9} 39.35%,
@@ -40,6 +39,13 @@ const HeroUnit = styled.div`
 `
 
 const ExploreButton = styled(Button)`
+  font-weight: normal;
+  color: white;
+  border: 2px solid white;
+  border-radius: 48px;
+  height: 60px;
+  cursor: pointer; /* until added to Button in ooni-components */
+
   &:hover, &:focus {
     background-color: white;
     color: ${props => props.theme.colors.blue5};
@@ -69,13 +75,22 @@ const StatsItem = ({label, unit, value }) => (
   <StyledStatsItem color='blue9' width={[1/3]} p={3}>
     <Text fontSize={[42, 48]} fontWeight={300}>
       {value}
-      <Text is='span' fontSize={32}>{unit}</Text>
+      <Text as='span' fontSize={32}>{unit}</Text>
     </Text>
     <StyledLabel>
       {label}
     </StyledLabel>
   </StyledStatsItem>
 )
+
+StatsItem.propTypes = {
+  label: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.string
+  ]),
+  unit: PropTypes.string,
+  value: PropTypes.number
+}
 
 const FeatureRow = styled(Flex)`
 
@@ -112,10 +127,6 @@ const ImgBox = styled.img`
   width: 100%;
 `
 
-const BorderedBox = styled(Box)`
-  border: 1px solid ${props => props.theme.colors.gray3};
-`
-
 const StyledContainer = styled(Container)`
   background-image: url('/static/images/world-dots.svg');
   background-repeat: no-repeat;
@@ -140,7 +151,8 @@ export default class LandingPage extends React.Component {
 
   render () {
     let {
-      measurementCount, asnCount,
+      measurementCount,
+      asnCount,
       countryCount
     } = this.props
 
@@ -160,7 +172,7 @@ export default class LandingPage extends React.Component {
                 <Text fontSize={[32, 64]} color='#ffffff'><FormattedMessage id='Home.Banner.Title.UncoverEvidence' /></Text>
               </Heading>
               <Text fontSize={[18, 24]} color='blue1'><FormattedMessage id='Home.Banner.Subtitle.ExploreCensorshipEvents' /></Text>
-              <ExploreButton mt={48} hollow inverted fontSize={24} onClick={() => (
+              <ExploreButton mt={48} px={5} hollow fontSize={24} onClick={() => (
                 Router.push('/countries')
               )}>
                 <FormattedMessage id='Home.Banner.Button.Explore' />
@@ -292,4 +304,10 @@ export default class LandingPage extends React.Component {
       </Layout>
     )
   }
+}
+
+LandingPage.propTypes = {
+  countryCount: PropTypes.number,
+  asnCount: PropTypes.number,
+  measurementCount: PropTypes.number
 }
